Add MIT license header
Make clear that the files I wrote, unlike the files I inherited, are under MIT
as opposed to OpD or PdroidManager's GPL. I think, when the tale is all told,
that the GPL will be what doomed PDroid (among several other things).

I don't see what use any of this might be for someone, but if they want it,
it is there for the taking.