Update reservoir for psx roms
I really only put in what I needed plus a little more for laziness. When
I finish auto-mating the auto-patcher I think I need to just add exactly the
files that fail. I could output the failed patches to a new file and read them
from there. Just run all my stored roms through and only pack what is essential.

No one is complaining and the program only needs to be downloaded once. But
it is larger than it needs to be at 100 mbs. There is a lot of fat that could
be trimmed.

Lets think about it.

closes #213