Update 4.3 provisionals with Slim-specific files
Must have deleted a patch that worked on those CDMAService files. Worth it to
turn down the program size so significantly. This was a bit of a pain, but
I added provisionals fromk a slim build instead of our usual default, which
is AOSP. Essentially, there was a patch in the slim patch sets that did not
apply cleanly to the AOSP CDMAServiceStateTracker provisional.

closes #259