Refine build query in regard to better VANIR check
The old function only checked for the presence of ro.goo.rom, for whatever reason. That meant that if
an omnirom (the only $ROM_check function after vanir) was built with goo integration, it was being slotted
as a vanir rom and never making it to the omni_check.

references #264