Fix breakages from ap_v3.7.9
Sorry guys, this was my fault alone. Made a bad assumption and ran with it.
Interested parties: I checked the source code patches for Camera.java and did
not see any changes in the Camera$Parameters or Camera$Events subclasses. So
I removed those smali patches from the autopatcher. Some of the static fields
(it turns out) are in use later in the file, so this caused issues once an
update zip was flashed. Thanks to @jz-train and @poo706 for the quick heads up
on the problem.

closes #275
closes #274