EXPIRATION function.
This is the "end-of-support" date. And it is  been set for OpenPDroid. I had intended to end support with Android 4.4. I feel we had gone too long without
a thorough audit, and I just don't have the time to do anything but maintain. Instead of stretching it out and watching the software work less and less, I wanted to
be proactive and end it ourselves. We worked on PDroid since Android 2.3, can you believe it?

Thanks for all the fish.

closes #289
closes #286
closes #285
closes #284
closes #283
closes #275
closes #282
closes #281
closes #277
closes #276
closes #270

I also added the requested EXPIRATION for tabletUI
refereneces #244
closes #244
closes #287